[Ps] Hunt harder in the model for the AGP version
Various new template forms place the AGP version in places other than a
classpath dependency: it is possible to declare the AGP version in a
plugins block of the main build file:

    plugins {
      id 'com.android.application' version '...' apply false
    }

or in the pluginManagement block of the settings file

    pluginManagement {
      plugins {
        id 'com.android.application' version '...'
      }
    }

As an extra complicating factor, in the case of the project-level build
file the versions must be literal strings, and cannot be variable
references (other than properties, though we don't support that at all
at the moment).  Therefore, when setting the AGP version, we need to be
able to set all the versions we've found at once.  In order to handle
this, construct in the parsedPropertyGetter lambda a synthetic
ResolvedPropertyModel which delegates to an arbitrary model for all
actions except setValue(), where it performs a setValue() on each of its
constituent models.

Fixes: 203665951
Test: manual
Change-Id: Ie0a587d71377e5f3450701aec63784c5d44e0a64