Tidy up and fix some issues with Android module icons
First we remove both AndroidStudioFileIconProvider and
ModuleNodeIconDecorator these are both superseded by
AndroidStudioIconProvider. The former uses FileIconPatcherImpl to
have FileIconProviders call into IconProviders for file-like objects.
The latter is a no-op since the icons are already set by the
IconProviders.

We move the functionality from AndroidStudioIconProvider to
AndroidIconProvider effectively undoing the changes from I56fc41ac3
apart from ensuring this code is only run against modules that come from
the Android plugin. This will cause the icons to be updated when run
from IDEA as well (this was being done by the ModuleNodeIconDecorator).

We update GradleUtil.getModuleIcon to ensure that the holder module and
main module display the icon based on the type of module, the unit test
module also displays the standard Java module icon and the Android test
displays the default Android app module icon.

We allow dynamic features to use the feature module icon. They already
use this in the new module wizard.

Bug: N/A
Test: visually verified
Change-Id: I344d17d29d16e102ed7872480c1b63676a618b94