Revert "Add a sourceset field to LintModelModuleLibrary"
This reverts commit 852d9a8f0e0bd197715c79349c2d007f0883012f.

Reason for revert: Moving from using a sourceset field to rely on the
artifact address to distinguish between different sourceset artifacts.
There is no need currently for lint to know whether a library or a
module is a testFixtures component of a library or a module. By
moving to using the artifact address lint will deal with them as
separate artifacts.

Change-Id: Iebd9569b46deebed75a17c7d60d75721bb3bfb04