Display and track invalid JDK reason on project
... open so it is consistent with what we do after syncing.

When a project is open we check if the Gradle JDK is valid and if not we
show an error, but we currently only show either "Could not determine
Gradle JDK" or "Could not find a valid JDK at <path>".

This change causes message in the error when opening a file to be the
same as the messages after syncing and also logs an event on
UsageTracker.

Bug: 173271723
Test: ProjectNotificationsUtilsTest
Change-Id: I7cda27a72b6fd7d13ecdd6e7671aeee45d092a42