Fix expected portion rendering when frame is only partly seen
When caching rectangles on model change, we used to clip the
ones partly visible. We now delay that and cache the full rectangles,
and only clip it during actual rendering.

This gives the rendering implementation the full rectangle size that
corresponds to the value, which is important for drawing content
within the rectangle that need to accurately visualize proportions.

Bug: n/a
Test: manual
Change-Id: Ia7c59cee44fc87a4a72580564f075f7643830784