Support preview in test modules
This change adds support for @Preview annotations present in android
test files, so that the Preview panel starts showing the @Preview
methods present in those files. The main changes needed for this
are to check whether the preview is in an android test file when
performing the build and when loading the classes.

Also, the test files are for now marked as a not valid location for
setting up a Compose Preview Run Configuration, and the @Preview
annotation is marked as not supported in unit test files.

Bug: 201895467
Test: tests added
Change-Id: I25f7c3ff71893310134cd21357bf5dbb0627ecdf