[BA] use same header for both parts of jetifier result presentation
The reason for the difference in headers on different systems that can
be observed on the screenshot provided in the bug comes from the fact
that these headers are not the same elements. Left part was a table
with a table header auto created but right part is a tree and it
does not have the same ui element used as the header.

Since left part does not actually use much of a table functionality,
we actually use it more as a list, replace table with a list view.
This allows us to control both columns' headers and make them look
as one on any system.

Fixes: 205008537
Test: updated existing
Change-Id: Idc1d79f4b91b210a93d2e2ed410b671c62054f1c