Add settings to disable startup agent for coroutine debugger
This change adds a setting to the Android Studio preferences
to allow the user to disable the coroutine debugger.

Disabling the coroutine debugger here means that, if disabled
when launching an app in debug mode the "attach_agent" command
is not passed to the "am start" command.

This is good to have in case something goes wrong in the agnet
and it prevents users from debugging their apps.

Test: CoroutinesDebuggerLaunchTaskContributorTest, CoroutineDebuggerConfigurableProviderTest
Bug: 182023904

Change-Id: Ia1e37f14a4f91da5cce303e555fdb60cc0b2f944