[AUA] Disable -SdkVersion operator rewrites
There are two problems with the current implementation.  Firstly, the
new operators are more restrictive in the set of string values they
accept; the old -SdkVersion properties would in some circumstances
accept strings of the form 'android-...', where the new ones don't, so a
simple operator renaming does not work in all situations (and it might
be a bit problematic to make it work by modifying a variable value,
which might be in use elsewhere in the build).

Secondly, the new pairs of operators are inconvenient for users who
might wish to switch between using a released Sdk and using a preview;
rather than being able to change a single variable, they will need to
change the operator being used in all their Gradle build files.  This
suggests that we might need to have another look at whether this is the
best solution to the problem of the -SdkVersion operators.

Bug: 201569704
Bug: 205806471
Test: existing tests adjusted
Change-Id: I96e84823747e9e278949a5ce8b429c8ebd16ced9
(cherry picked from commit 9ae1bf803f81b0507df89b6a0b2d326d076e6bf1)