Remove expensive call from update method
This call is not required, the validation it check happens when the
action is pressed. This also removes user confusion of when the action
is avaiable.

We also remove a code path that throws an exception when the
GradleBuildModel is un-available and replace it was a dialog letting the
user know that this action can't be completed automatically.

Bug: N/A
Test: N/A
Change-Id: I5615ec137132faa1c99fe7c6951eb54394e3c801