Create a Special Token For Filtering Project Apps
This feature was suggested by Tor during our 1-1. The idea is to make it easy to write a filter that matches the app under development.

This is useful when we need to filer our app OR some other condition, for example, show only our app or any error.

   "app! | level:E"

In order for this to work, I had to do a significant refactoring of the LogcatFilter class hierarchy. The underlying reason for the refactor is that when the ProjectAppFilter (renamed) is used in a combined filter as opposed to added as a top level filter, it needs to be able to reload the set of project apps. We still want to do this only once per batch so we have to introduce a "prepare()" method to the LogcatFilter interface and call in before each batch.

Bug: 191880898
Test: Added
Change-Id: I59c77cc65fd9b9276fd86a12004b511f1bba9d91