[AUA] Move GradlePluginUpgradeState to sync module
The GradlePluginUpgradeState is a nexus for two pieces of policy:
whether a particular combination of project and current AGP versions
should be considered compatible enough to even attempt sync; and whether
a combination of project and currently-available AGP versions should
trigger an upgrade recommendation.  For the first part of this
computation to be effective, it needs to be available to sync and in
particular to the Gradle process, with its stringent restrictions on the
run-time classpath.  Having the second part (the IDE policy on
recommendations) might not be quite right, but it's not utterly wrong either.

Bug: 206397060
Test: existing
Change-Id: I7d45880be680d582a06ce758f1cc3d9605cdc966