Fix NPE in testApi29VersionBanner
The AppInspectionApiServices instance was specified as a
mockito mock when creating a AppInspectionInspectorClient
instance (3 different places in the test).

The test itself did not fail, but the NPE would happen
after the test finished during the actual connect where
AppInspectionApiServices.attachToProcess would return null.

The subsequent call to getLibraryVersions will cause a NPE.

Fixes: 204343003
Test: testApi29VersionBanner
Change-Id: Ia2fad11b7a2c022910dff8dc1a141a6f72021a5e