[Sync] Allow user-overrides of forced upgrades
Preserve the logic from shouldForcePluginUpgrade to the new
forced-upgrade detection in sync by propagating it to the Gradle process
as a boolean; for naming purposes, pretend that it is purely a flag,
though the logic is slightly more involved than that.  Use that boolean
when deciding whether to force the upgrade.

Bug: 206397060
Bug: 206666447
Test: none
Change-Id: I12f92a510a13632208c3c64f02d4b634db21cc12