[Dsl] New appliedPlugins() model method
Until the new forms of plugin specification, all plugins were applied to
the current project.  The new forms allow specifying `apply false`, to
indicate that the current project should have this plugin available but
not necessarily applied unless specified.  Define a new appliedPlugins()
method which returns PluginModels for only those plugins that are in
fact applied in the particular context, and implement it for
GradleBuildModel (which collects plugins from the various application
syntaxes) and for PluginsBlockModel (which is used only in
settings pluginManagement).

Bug: none
Test: existing tests extended
Change-Id: I77e247e470809e040aef5a952a801e4ad61bb128