[Ps] Move a test
Remove the placeholder test in projectSystem.gradle.psd.tests, and
replace it with PsModuleTypeTest.  Provide test data and dependencies
cribbed from projectSystem.gradle.tests, on the basis that we will end
up using the same projects from android/testData/projects in the Project
Structure / PSD tests, even if not all of them are used in the one test
that has moved.

Bug: n/a
Test: existing
Change-Id: I3dc441255d040369bdadb90821d80af0dc9e23fe