[BA] Make "last updated" string change with time while page is open
Use Alarm to make ui auto-refresh every 30 seconds to update strings
that are time-dependant (e.g. last run time format in this case).
There might also be other things to refresh in the future,
for example we can periodically check if build files where changes
and show results outdated banner in this case.

Bug: 206824616
Test: N/A
Change-Id: Ib59e00ec6723b5123904f5f18b0711b4ee077041