Add tests for ComposePreviewView
This tests the interaction with the view component of the Compose
preview.
This ensures that the contents are displayed and that the empty and
error states are correctly handled.

This change also updates WorkBench to allow a delay time of 0 for
refresh, meaning that the loading state will be immediately visible
after being triggered by startLoading. This is what we want for the
Compose Preview as opposed to waiting 1 seconds to start it.

Bug: N/A
Test: Adds new test
Change-Id: I7db823a1b497f23a7e053ae939e20249c0a8d2e2
(cherry picked from commit 56fe82ccd72771f3eb9f76a548813669befac948)