Change Logging Style of 'beginning of...' Messages
When logcat starts or is cleared, it emits lines like:
--------- beginning of kernel
--------- beginning of system

It seemed like a good idea to log inject a log entry for these lines but as it turns out, if the device time is different than the host time, the fake lines will have a different timestamp making it seem like the logs are out of order.

So, instead of logging them as normal messages, we'll log them the same way they are emitted by logcat utility itself, without a header.

The easiest way to mark them as system messages is to use a timestamp of Instant.EPOCH as a marker.

Bug: 191880898
Test: Added
Change-Id: Ic647f902f539f9ef9f2582b386a7b03dafe58d12