Remove myDebuggeeModuleProvider from AndroidDebuggerState
As we are trying untangle Debuggers from anything but Clients and debugger
setting I'm cleaning up AndroidDebuggerState.
AndroidDebuggerState should contain only debugger pure settings,
because we reuse this state between different modules/apps.
We still need facet, but I hope to remove it in future CLs.

The myDebuggeeModuleProvider is equal to the module in the facet,
 because facet received via AndroidFacet.getInstance(module).
See AndroidRunConfigurationBase.createLaunchTasksProvider.

Bug: 210440201
Test: refactoring
Change-Id: I0b5437cfbc714bc658e9c67c2d88520fb82f8a47