[AUA] Adjust from version for some migrations
There was an issue with the receiver type for some Groovy
implementations of the new blocks, manifesting as a failure to find an
applicable method in Arctic Fox / 7.0.0.  This was fixed in the 7.0.1
patch release, so only offer those migrations for that version or
higher.

Bug: 194937528
Test: existing tests adjusted
Change-Id: Ie51dce4abdc7f658910becfd1ab83b1cbe2e2c17
(cherry picked from commit 7a9d5162687e0199301853672182ed6bdc77e558)