Remove Infer Support Annotations action
This is an experimental action which has never been
enabled in Studio (though it was described at I/O
and there were some docs on it on tools.android.com).

It was not using the studio flags mechanism to gate
access (it predates that facility) and apparently at
some its attempts to disable itself stopped working.

Test: N/A
Bug: Not filed?
Change-Id: Ia71d306d79455de0edb517bcb55a607b2dd96718