Fix artifact when repainting last rectangle in StateChart
When the rectangle to redraw happens to be the last one in StateChart,
the handling wasn't right. The "left" and "right" indices surrounding
the mouse were identical (due to clamping), and in some cases we ended
up with `modelXRight` being to the left of `modelXLeft` (because we
took `min`/`max` with the range's upper/lower bound, respectively),
resulting a negative rectangle width, essentially telling Swing not
to redraw under the mouse.

This CL refactors `seriesIndexAtMouse` and fixes the edge cases.

Fixes: 200817402
Test: add new
Change-Id: I99400e620174d646e58cb4197619917751fa6544