Fix flaky test due to double initialization
testDownloadFileLocationWithMouseClick is the only flaky test of the
suite, and is also the only one to call downloadFile() twice. The flake
occurs in the second downloadFile() operation.

downloadFile() performs some setup operations, such as creating the
controller. Thus, on the second invocation, there are two controllers
associated with the same view and model. I have not tracked down exactly
how this causes a flake, but it does not seem like a desirable test
configuration.

Extracting controller setup to downloadFileSetup() resolves the flake.
(This could potentially be consolidated with many other cases that
repeat the same four lines, though I wanted to keep this diff small.)

Fixes: 203403492
Test: checked with runs_per_test=200
Change-Id: Id7c7b4beb7d43651faf60f6eee5853e4444f7623