LE: Support functions has param / return type of Composables.
Given a compose function, we remap the function's parameters by adding
state variables that the compose compiler inserts.

However, we currently don't do that for cases where the parameters
or return types are also composables.

This CL fixes the issues by doing the mapping on the parameters
as well as the return type.

We do NOT, however, need to do this recursively because of
type erasure.

For example, a composable parameter be mapped from Function2 to
Function4. The types of the parameters are lost because of
type erasure.

Test: BasicCompileTest.kt
Bug: none
Change-Id: I57cb187e6c42deb52052d3b9c782469a69bc4a6d