Do not publish facetConfigurationChanged() when attaching models
`facetConfigurationChanged` notification is consumed by
`AndroidDependenciesCache` and `SourceProviderManagerComponent` only.

Publishing this even multiple times (thousands of times for large
projects) during Gradle sync causes poor sync performance.

`AndroidDependenciesCache` it is reset on project root changed
events via `AndroidProjectRootListener`, which is published every time
the structure of the IDE project changes regardless of the project
system in use.

Itself it does not depend on models and does not need to be refreshed
when they are loaded asynchronously.

`SourceProviderManagerComponent` uses these events to detect the moment
when a project becomes an Android project and at the moment Gradle sync
runs it is already an Android project.

Bug: n/a
Test: n/a (existing)
Change-Id: Id661bca4b6f338ba0932c95e4d391bd8188743ff