Streamline LiveEdit concurrency to prevent IDE lag
The LiveEdit change detection & handling flow is as follows:

There are three thread contexts:
 - The UI thread, which reports PSI events
 - The LiveEditService executor, which queues changes and schedules
   LiveEdit pushes. Single-threaded.
 - The AndroidLiveEditDeployMonitor executor, which handles
   compile/push of LiveEdit changes. Single-threaded.

LiveEditService registers a single PSI listener with the PsiManager.
This listener receives callbacks on the UI thread when PSI
events are generated. There is one LiveEditService instance per Project.

AndroidLiveEditDeployMonitor registers one LiveEditService.EditListener
per Project with the corresponding Project's LiveEditService. When the
LiveEditService receives PSI events, the listener receives a callback
on a single-threaded application thread pool owned by the
LiveEditService.

The EditListener callback enqueues the event in a collection of
"unhandled" events, schedules a LiveEdit compile+push, and returns
quickly to allow the thread pool to continue enqueuing events.

The scheduled LiveEdit compile+push is executed on a single-threaded
executor owned by the EditListener. It handles changes as follows:
 1. Lock the queue  of unhandled changes
 2. Make a copy of the queue, clear the queue, then unlock the queue
 3. If the copy is empty, return
 4. Attempt to compile and push the copied changes
 5. If the compile is successful, return.
 6. If the compile is cancelled, lock queue, read-add the removed
 events, then schedule another compile+push

 Compilation may be cancelled by PSI write actions, such as the user
 continuing to type after making a change. It may also be prevented by
 an ongoing write action, or a PSI write action from another source,
 which is why it is safer to schedule a retry rather than assuming
 whatever PSI modification cancelled the change will cause a LiveEdit
 push.

 Note that this retry logic does NOT apply if the compilation explicitly
 *fails*; only if it is cancelled by PSI write actions. This is why the
 queue is cleared before attempting compilation, as a failed compile has
 still "handled" the changes.

 Compilation is responsible for handling duplicate changes
 originating from the same file, and performs de-duplication logic to
 ensure that the same file is not re-compiled multiple times.

Bug: None
Test: Manual
Change-Id: Ifa52a517d32498e2c087fa9746b1a9dc6adc3a56