Fix loading of R ids from non transitive R classes
Project library dependencies were being considered as regular external
libraries and its ids treated as non-final. This would cause the
ViewLoader to dynamically generate ids and this would break the preview
as the ids are inlined and can not be dynamic.

This change ensures that the R class is loaded from its compiled version
for project library dependencies so dynamic ids are not generated.

Fixes: 206862224
Test: NonTransitiveResourcesLoaderTest
Change-Id: Ide05aed92bcddb3aa04a715017fe486731b4a895