[AUA] Add an ActionLink for more information
The visible link is better than a fairly non-discoverable tooltip over
the "Run selected steps" button (though keep that for now as well).
However, components capable of displaying a link (and responding to
the action event) have a different text baseline relative to their
overall extent, so in order for the link not to be jarringly ugly next
to the plain text status message we need to use a layout
which (mostly) respects text baselines.

Bug: 213019734
Test: manual
Change-Id: I4251253abb0aa18a979794642a36f778534fc85c