[ResManager] Exlusively use main modules on the entire tool
Not sure why but source sets seem to be exposed as modules now, which
doesn't make sense from a resource management perspective.

This filters out all redundant modules in favor of the `main` module.
Reasoning is that the main module is the only api to differentiate the
hierachy of these modules.

Fixes: 204445192
Test: n/a, unit tests don't reflect production module hierarchy, this is
not a fix, but rather a mitigation

Change-Id: I5965a1143ef36f5648024e71cc55439f3df71881