Prevent unnecessary refresh when project needs to build
We were triggering the entire Compose Preview refresh flow even when
the project was not built yet. In addition to waste resources when in
cases that would not actually render the preview, we were breaking the
state of our refresh model.

An example of a side effect is we were not zooming the preview to fit
after rendering it for the first time. The reason is we were calling
"onAfterRender" *before* rendering the preview for the first time, as
a result of the unnecessary refresh triggered when opening the file
with the empty (i.e. "Project needs to build") panel.

Fixes: 192832629
Bug: 215190504
Test: Manual by creating a new project and building once.
Change-Id: I14d86961de16dd299cca10f7e430bd6d332efdee