Fix Flow helper DnD issue
A issue was reported that layout_editor_absoluteX and layout_editor_absoluteY of a widget should be removed when this widget is grouped into a Flow helper since both attributes are no longer valid.

In this CL, changes are made to remove both attributes when a widget is dragged and dropped into a Flow helper in the Component Tree Panel.

Bug: 160716293
Test: Manually tested
Change-Id: Ibced3019ba2c85e4ccd42632a46e0b7addd4253d