Reduce compilation time for Compose tests
This changes the compilation task triggered for tests from assembleDebug
to compileDebugSources.
It also removes an unnecessary initial compilation for
LiveLiteralsAndLiveEditIntegrationTest and moves the work from the
limited io pool to worker threads.

This change also disables one integration test in
LiveLiteralsAndLiveEditIntegrationTest since it seems to be flaky for
unrelated reasons.

Bug: 161091273
Test: bazel test --runs_per_test=50 //tools/adt/idea/compose-designer:intellij.android.compose-designer.tests_tests__gradle
Change-Id: I5ae17ed4aac30a7ae1a2aa10b96dd3a96a72eaad